Luke 15, del 1

+++ 2023/src/_december_template.rs
@@ -1,9 +1,9 @@
-pub fn part_1(input: &str) -> i16 {
+pub fn part_1(input: &str) -> u32 {
 
     0
 }
 
-pub fn part_2(input: &str) -> i16 {
+pub fn part_2(input: &str) -> u32 {
 
     0
 }
